+++ lib/decorators/get.ts
@@ -12,6 +12,6 @@ import { requestDecoratorsFactory } from './helpers/request-decorators.factory';
 /**
  * Creates handler which listen GET requests
  */
-export function GET(config: RouteConfig) {
+export function GET(config: string | RouteConfig) {
     return requestDecoratorsFactory('get')(config);
 }