+++ lib/decorators/head.ts
@@ -12,6 +12,6 @@ import { requestDecoratorsFactory } from './helpers/request-decorators.factory';
 /**
  * Creates handler which listen HEAD requests
  */
-export function HEAD(config: RouteConfig) {
+export function HEAD(config: string | RouteConfig) {
     return requestDecoratorsFactory('head')(config);
 }