+++ src/handlers/post.handler.ts
@@ -9,11 +9,23 @@
 import { POST, RequestHandler } from 'fastify-decorators';
 
 @POST({
-    url: '/post'
+    url: '/post',
+    options: {
+        schema: {
+            response: {
+                200: {
+                    type: 'object',
+                    properties: {
+                        message: { type: 'string' }
+                    }
+                }
+            }
+        }
+    }
 })
 class PostHandler extends RequestHandler {
     public async handle(): Promise<any> {
-        return {message: this.request.body.message};
+        return this.request.body;
     }
 }